refactor(adapter-openclaw): extract approval converters into approval-helpers.ts
Sixth cut of the runtime.ts god-file split. Moves the 9 pure approval
converters — renderNativeApprovalDescription, supportsNativeApprovalOptions,
requiresRejectReason, approvalEventFromOpenClawPayload, channelIdFromOpenClawRequest,
openClawDecisionFromBakinOption, bakinOptionFromOpenClawDecision,
parseNativeApprovalRef, isExpectedNativeApprovalResolveMiss — plus their
constants (OPENCLAW_PLUGIN_ID, OPENCLAW_WORKFLOW_GATE_TOOL,
OPENCLAW_PLUGIN_APPROVAL_REF_PREFIX, NATIVE_APPROVAL_NOTICE) into a new
approval-helpers.ts. Pure mapping between Bakin approval events and OpenClaw
plugin-approval payloads/decisions/refs; types come from ./approval-gateway +
core, no adapter state.

approvalNoticeForMessage stays in runtime.ts: it reads channel-approval
capability (channelHasInteractiveApproval) + the RENDER_ONLY/REJECT_REASON
notices, so it travels with the channel cluster in a later cut, not here.
The class's approval methods import back the converters + 3 shared constants
they use; the rest stay module-internal.

Pure relocation — bodies byte-for-byte unchanged. runtime.ts 2,241 -> 2,156.
Verified: typecheck, eslint (changed files), adapter suite 115-0 incl.
runtime-channels.test.ts (--isolate), full suite 5124-0, binary build (stamps
reverted), boot smoke (adapter loads, schedule activates, list route 200).
